Refactored src files for easier importing
