--- src/LBFGS/__init__.py
@@ -1,9 +1,0 @@
-from .LBFGS import FullBatchLBFGS
-from . import LBFGS
-from . import utils
-
-__all__ = [
-    "FullBatchLBFGS",
-    "LBFGS",
-    "utils",
-]
