+++ wenda_gpu/LBFGS.py
@@ -4,7 +4,7 @@ import matplotlib.pyplot as plt
 from functools import reduce
 from copy import deepcopy
 from torch.optim import Optimizer
-
+from wenda_gpu import LBFGS_utils
 
 def is_legal(v):
     """
