+++ wenda_gpu/LBFGS_utils.py
